Added tests and documentation
